Delete README.md
Objective 
To find the best entertainer from the given data set .

Method
I choice 3 methods to find the best entertainer.
1>	Who gets famous at what age.
2>	Who is able to maintain his dominance longer in the industry.
3>	Which year lost a major role in the industry.

Data processing 
There are 3 different data set available in the CSV format namely :
Entertainer =Basic info, Entertainer – Breakthrough info, Entertainer – Last work info.
In the beginning, I downloaded the all file from the given link on the laptop and opened in the tableau desktop. The very next step involves the data joining (inner join) as I want to perform the analysis on those data only which matches the 3 CSV file. Here I found 1 candidate key i.e. Name. which was common in all three data sets. Even checked the entertainers are alive or dead .

Conclusion :

Here found Stevie wonder is the youngest entertainer to receive fame at the age of 13 only. with Morgan Freeman with eldest at age of 63. Secondly between 50-55 is the maximum number of years of dominance by the entertainer with a total number of 9 . Lastly Found the between 2015-2020 is the year where most of the entertainers have lost their fame which in the count is 31.

LINK:  To  Preview Dashboard 
https://public.tableau.com/app/profile/sushant3538/viz/Amazonsellinjune/Dashboard1